+++ src/main/scala/PspTypes.scala
@@ -3,7 +3,7 @@ package core
 
 /** A thin abstraction over some questionable assumptions. */
 trait PspTypes {
-  type Index         = Int
+  type Index         = Long
   type Done          = Boolean
   type Suspended[+A] = (A => Unit) => Unit
   type Ref[+A]       = A with AnyRef
